Reorganize CopilotStudio client files into agent-to-agent dotnet directory structure
