--- samples/basic/CopilotStudioClient.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Worker">
-
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-		<Nullable>enable</Nullable>
-		<ImplicitUsings>enable</ImplicitUsings>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.1" />
-		<PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.67.2" />
-		<PackageReference Include="Microsoft.Agents.CopilotStudio.Client" Version="1.2.*-*"/>
-	</ItemGroup>
-
-</Project>
