+++ samples/basic/copilotatudio-agentToagent/dotnet/appsettings.json
@@ -1,0 +1,26 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Error",
+      "Microsoft.Hosting.Lifetime": "Warning"
+    }
+  },
+  "CopilotStudioClientSettings1": {
+    "DirectConnectUrl": "",
+    "EnvironmentId": "b4ccc464-f5b7-e266-bada-06122c419e03", 
+    "SchemaName": "cr73a_testAgent", 
+    "TenantId": "391bca38-63d1-4cbd-ba78-c74369b91888",
+    "UseS2SConnection": false,
+    "AppClientId": "525dc2dd-a78d-4770-9c21-71bea65fab56",
+    "AppClientSecret": ""
+  },
+  "CopilotStudioClientSettings2": {
+    "DirectConnectUrl": "",
+    "EnvironmentId": "a865c82b-3806-efa9-a5a3-df5889dfa013", 
+    "SchemaName": "cr082_agent2", 
+    "TenantId": "391bca38-63d1-4cbd-ba78-c74369b91888",
+    "UseS2SConnection": false,
+    "AppClientId": "525dc2dd-a78d-4770-9c21-71bea65fab56",
+    "AppClientSecret": ""
+  }   
+}
