+++ samples/basic/copilotatudio-agentToagent/dotnet/dotnet.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CopilotStudioClient", "CopilotStudioClient.csproj", "{207DA8CF-1247-A6F5-60AA-6C634C5E5BF3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{207DA8CF-1247-A6F5-60AA-6C634C5E5BF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{207DA8CF-1247-A6F5-60AA-6C634C5E5BF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{207DA8CF-1247-A6F5-60AA-6C634C5E5BF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{207DA8CF-1247-A6F5-60AA-6C634C5E5BF3}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8713DD82-C2D9-41B9-B2D7-480AB84702E5}
+	EndGlobalSection
+EndGlobal
